refactor shapes serializers
