+++ nanome/_internal/shapes/__init__.py
@@ -4,5 +4,5 @@ from .line import _Line
 from .label import _Label
 from .sphere import _Sphere
 from .mesh import _Mesh
-from . import serialization
+from . import serializers
 from . import _deprecated
