--- nanome/_internal/shapes/serialization/__init__.py
@@ -1,6 +1,0 @@
-from .anchor_serializer import _AnchorSerializer
-from .sphere_serializer import _SphereSerializer
-from .line_serializer import _LineSerializer
-from .label_serializer import _LabelSerializer
-from .mesh_serializer import _MeshSerializer
-from .shape_serializer import _ShapeSerializer
