--- nanome/_internal/shapes/serialization/anchor_serializer.py
@@ -1,30 +1,0 @@
-from nanome._internal.util.type_serializers import TypeSerializer, UnityPositionSerializer
-from nanome._internal.shapes import _Anchor
-
-
-class _AnchorSerializer(TypeSerializer):
-    def __init__(self):
-        self._offset = UnityPositionSerializer()
-
-    def version(self):
-        return 0
-
-    def name(self):
-        return "ShapeAnchor"
-
-    def serialize(self, version, value, context):
-        context.write_long(value._target)
-        context.write_byte(int(value._anchor_type))
-        context.write_using_serializer(self._offset, value._local_offset)
-        context.write_using_serializer(self._offset, value._global_offset)
-        context.write_using_serializer(self._offset, value._viewer_offset)
-
-    def deserialize(self, version, context):
-        from nanome.util.enums import ShapeType
-        result = _Anchor._create()
-        result._target = context.read_long()
-        result._anchor_type = ShapeType.safe_cast(context.read_byte())
-        result._local_offset = context.read_using_serializer(self._offset)
-        result._global_offset = context.read_using_serializer(self._offset)
-        result._viewer_offset = context.read_using_serializer(self._offset)
-        return result
