+++ nanome/_internal/shapes/serializers.py
@@ -1,0 +1,222 @@
+from nanome._internal.shapes import _Anchor
+from nanome._internal.shapes.label import _Label
+from nanome._internal.shapes.line import _Line
+from nanome._internal.shapes.mesh import _Mesh
+from nanome._internal.shapes.sphere import _Sphere
+from nanome._internal.util.type_serializers import TypeSerializer
+from nanome._internal.util.type_serializers import TypeSerializer, StringSerializer
+from nanome._internal.util.type_serializers import TypeSerializer, UnityPositionSerializer
+from nanome._internal.util.type_serializers import TypeSerializer, UnityPositionSerializer, ColorSerializer, ArraySerializer
+import os
+import tempfile
+
+
+class _AnchorSerializer(TypeSerializer):
+    def __init__(self):
+        self._offset = UnityPositionSerializer()
+
+    def version(self):
+        return 0
+
+    def name(self):
+        return "ShapeAnchor"
+
+    def serialize(self, version, value, context):
+        context.write_long(value._target)
+        context.write_byte(int(value._anchor_type))
+        context.write_using_serializer(self._offset, value._local_offset)
+        context.write_using_serializer(self._offset, value._global_offset)
+        context.write_using_serializer(self._offset, value._viewer_offset)
+
+    def deserialize(self, version, context):
+        from nanome.util.enums import ShapeType
+        result = _Anchor._create()
+        result._target = context.read_long()
+        result._anchor_type = ShapeType.safe_cast(context.read_byte())
+        result._local_offset = context.read_using_serializer(self._offset)
+        result._global_offset = context.read_using_serializer(self._offset)
+        result._viewer_offset = context.read_using_serializer(self._offset)
+        return result
+
+
+class _LabelSerializer(TypeSerializer):
+    def __init__(self):
+        self._string = StringSerializer()
+
+    def version(self):
+        return 0
+
+    def name(self):
+        return "LabelShape"
+
+    def serialize(self, version, value, context):
+        context.write_using_serializer(self._string, value._text)
+        context.write_float(value._font_size)
+
+    def deserialize(self, version, context):
+        result = _Label._create()
+        result._text = context.read_using_serializer(self._string)
+        result._font_size = context.read_float()
+        return result
+
+
+class _LineSerializer(TypeSerializer):
+    def __init__(self):
+        pass
+
+    def version(self):
+        return 0
+
+    def name(self):
+        return "LineShape"
+
+    def serialize(self, version, value, context):
+        context.write_float(value._thickness)
+        context.write_float(value._dash_length)
+        context.write_float(value._dash_distance)
+
+    def deserialize(self, version, context):
+        result = _Line._create()
+        result._thickness = context.read_float()
+        result._dash_length = context.read_float()
+        result._dash_distance = context.read_float()
+        return result
+
+
+class _MeshSerializer(TypeSerializer):
+    def __init__(self):
+        pass
+
+    def version(self):
+        return 1
+
+    def name(self):
+        return "MeshShape"
+
+    def read_texture(self, value):
+        from nanome.util import Logs
+        if value.texture_path != "":
+            filename, ext = os.path.splitext(value.texture_path)
+            if ext.lower() in [".jpeg", ".jpg", ".png"]:
+                path = value.texture_path.replace("\\", "/")
+                if os.path.isfile(path):
+                    try:
+                        with open(path, "rb") as f:
+                            texture_bytes = bytearray(f.read())
+                            return texture_bytes
+                    except Exception as e:
+                        Logs.error("Error reading texture file: " + e)
+                else:
+                    Logs.error("Texture file does not exist")
+            else:
+                Logs.error("Texture file should be a png or a jpg file")
+        return []
+
+    def serialize(self, version, value, context):
+        from nanome.util import Logs
+        context.write_float_array(value.vertices)
+        context.write_float_array(value.normals)
+        context.write_float_array(value.colors)
+        context.write_int_array(value.triangles)
+        context.write_float_array(value.uv)
+
+        texture_bytes = self.read_texture(value)
+        context.write_byte_array(texture_bytes)
+        if len(texture_bytes) > 0:
+            Logs.debug("Sending texture:", value.texture_path)
+
+        if version >= 1:
+            context.write_bool(value.unlit)
+
+    def create_texture_file(self, texture_path, texture_bytes):
+        with open(texture_path, "wb") as f:
+            f.write(texture_bytes)
+
+    def deserialize(self, version, context):
+        result = _Mesh._create()
+        result.vertices = context.read_float_array()
+        result.normals = context.read_float_array()
+        result.colors = context.read_float_array()
+        result.triangles = context.read_int_array()
+        result.uv = context.read_float_array()
+        texture_bytes = context.read_byte_array()
+
+        if version >= 1:
+            result.unlit = context.read_bool()
+
+        if len(texture_bytes) > 0:
+            temp_texture = tempfile.NamedTemporaryFile(
+                delete=False, suffix='png')
+            self.create_texture_file(temp_texture.name, texture_bytes)
+            result.texture_path = temp_texture.name
+        return result
+
+
+class _ShapeSerializer(TypeSerializer):
+    def __init__(self):
+        self._position = UnityPositionSerializer()
+        self._color = ColorSerializer()
+        self._sphere = _SphereSerializer()
+        self._line = _LineSerializer()
+        self._label = _LabelSerializer()
+        self._mesh = _MeshSerializer()
+        self._anchor_array = ArraySerializer()
+        self._anchor_array.set_type(_AnchorSerializer())
+
+    def version(self):
+        return 1
+
+    def name(self):
+        return "Shape"
+
+    def serialize(self, version, value, context):
+        from nanome.util.enums import ShapeType
+        context.write_byte(int(value._shape_type))
+        if value.shape_type == ShapeType.Sphere:
+            context.write_using_serializer(self._sphere, value)
+        elif value.shape_type == ShapeType.Line:
+            context.write_using_serializer(self._line, value)
+        elif value.shape_type == ShapeType.Label:
+            context.write_using_serializer(self._label, value)
+        elif value.shape_type == ShapeType.Mesh:
+            context.write_using_serializer(self._mesh, value)
+        context.write_int(value._index)
+        context.write_using_serializer(self._anchor_array, value._anchors)
+        context.write_using_serializer(self._color, value._color)
+
+    def deserialize(self, version, context):
+        from nanome.util.enums import ShapeType
+        shapeType = ShapeType.safe_cast(context.read_byte())
+        result = None
+        if shapeType == ShapeType.Sphere:
+            result = context.read_using_serializer(self._sphere)
+        elif shapeType == ShapeType.Line:
+            result = context.read_using_serializer(self._line)
+        elif shapeType == ShapeType.Label:
+            result = context.read_using_serializer(self._label)
+        elif shapeType == ShapeType.Mesh:
+            result = context.read_using_serializer(self._mesh)
+        result._index = context.read_int()
+        result._anchors = context.read_using_serializer(self._anchor_array)
+        result._color = context.read_using_serializer(self._color)
+
+        return (context.read_int(), context.read_bool())
+
+
+class _SphereSerializer(TypeSerializer):
+    def __init__(self):
+        pass
+
+    def version(self):
+        return 0
+
+    def name(self):
+        return "SphereShape"
+
+    def serialize(self, version, value, context):
+        context.write_float(value.radius)
+
+    def deserialize(self, version, context):
+        result = _Sphere._create()
+        result.radius = context.read_float()
+        return result
